Rename Finnhub file to be more generic
