+++ server/src/middleware/validateTicker.ts
@@ -1,7 +1,7 @@
 import express from 'express';
 
 import { BadRequestError } from '../errors';
-import { StockTransactionAPI } from '../api/finnhub';
+import { StockTransactionAPI } from '../api/stockTransactionAPI';
 
 /** Checks if a stock ticker is included in the request and if it's a valid
  * ticker on the NYSE (according to Finnhub)
